+++ app/middleware/cleanstate.js
@@ -1,0 +1,16 @@
+exports = module.exports = function(store) {
+  
+  return function(options) {
+    options = options || {};
+    options.store = store;
+    
+    return require('flowstate').clean(options);
+  };
+};
+
+exports['@implements'] = [
+  'http://i.bixbyjs.org/http/middleware/cleanState'
+];
+exports['@require'] = [
+  '../state/store'
+];